added in new ui class, adds table to track results, adds trinket parsing from SU data, upgrades dash 2.0.0+, as copy/paste of SU data
