+++ Changes.md
@@ -1,0 +1,5 @@
+* Added abilty to paste in json from Seventy Upgrades
+* Added ability to parse trinkets from Seventy Upgrades data
+  * Assumes trinkets from Seventy Upgrades unless explicitly picked from dropdowns
+* Added sim results table to track results
+* upgrades dash to 2.0.0+
