+++ README.md
@@ -2,3 +2,11 @@
 Full rotation simulator for feral DPS in WoW TBC Classic.
 
 The file tbc_cat_sim.py contains all of the inner workings of the sim, including how character resources and actions are modeled, how attack tables are executed, and the simulation rotation logic. The file main.py contains the GUI used for the web application. The first file is well documented as it is the important part of the sim, and the second file is a bit of a mess currently.
+
+
+## Docker
+```shell script
+docker build -t catsim .
+docker run -it -p 8080:8080 catsim
+
+```
