+++ requirements.txt
@@ -1,10 +1,10 @@
 Brotli==1.0.9
 click==7.1.2
-dash==1.19.0
-dash-core-components==1.15.0
-dash-html-components==1.1.2
+dash==2.1.0
+dash-core-components>=2.0.0
+dash-html-components==2.0.0
 dash-renderer==1.9.0
-dash-table==4.11.2
+dash-table>=5.0.0
 Flask==1.1.2
 Flask-Compress==1.9.0
 future==0.18.2
@@ -13,9 +13,10 @@ itsdangerous==1.1.0
 Jinja2==2.11.3
 MarkupSafe==1.1.1
 numpy==1.19.5
-plotly==4.14.3
+plotly>=5.0.0
 retrying==1.3.3
 six==1.15.0
 Werkzeug==1.0.1
 dash-bootstrap-components==0.12.0
 psutil==5.8.0
+pandas==1.4.2
