+++ ui/outputs.py
@@ -1,0 +1,236 @@
+from dash import dash_table, dcc, html
+import dash_bootstrap_components as dbc
+
+sim_output = dbc.Col([
+    html.H4('Results'),
+    dcc.Loading(children=html.Div([
+        html.Div(
+            'Average DPS:',
+            style={
+                'width': '50%', 'display': 'inline-block',
+                'fontWeight': 'bold', 'fontSize': 'large'
+            }
+        ),
+        html.Div(
+            '',
+            style={
+                'width': '50%', 'display': 'inline-block', 'fontSize': 'large'
+            },
+            id='mean_std_dps'
+        ),
+    ]), id='loading_1', type='default'),
+    dcc.Loading(children=html.Div([
+        html.Div(
+            'Median DPS:',
+            style={
+                'width': '50%', 'display': 'inline-block',
+                'fontWeight': 'bold', 'fontSize': 'large'
+            }
+        ),
+        html.Div(
+            '',
+            style={
+                'width': '50%', 'display': 'inline-block', 'fontSize': 'large'
+            },
+            id='median_dps'
+        ),
+    ]), id='loading_2', type='default'),
+    dcc.Loading(children=html.Div([
+        html.Div(
+            'Time to oom:',
+            style={
+                'width': '50%', 'display': 'inline-block',
+                'fontWeight': 'bold', 'fontSize': 'large'
+            }
+        ),
+        html.Div(
+            '',
+            style={
+                'width': '50%', 'display': 'inline-block', 'fontSize': 'large'
+            },
+            id='time_to_oom'
+        ),
+    ]), id='loading_oom_time', type='default'),
+    html.Br(),
+    html.H5('DPS Breakdown'),
+    dcc.Loading(children=dbc.Table([
+        html.Thead(html.Tr([
+            html.Th('Ability'), html.Th('Number of Casts'), html.Th('CPM'),
+            html.Th('Damage per Cast'), html.Th('DPS Contribution')
+        ])),
+        html.Tbody(id='dps_breakdown_table')
+    ]), id='loading_3', type='default'),
+    html.Br(),
+    html.H5('Aura Statistics'),
+    dcc.Loading(children=dbc.Table([
+        html.Thead(html.Tr([
+            html.Th('Aura Name'), html.Th('Number of Procs'),
+            html.Th('Average Uptime'),
+        ])),
+        html.Tbody(id='aura_breakdown_table')
+    ]), id='loading_auras', type='default'),
+    html.Br(),
+    html.Br()
+], style={'marginLeft': '2.5%', 'marginBottom': '2.5%'}, width=4, xl=3)
+
+## stats output
+
+stats_output = dbc.Col(
+    [html.H4('Raid Buffed Stats'),
+     html.Div([
+         html.Div(
+             'Swing Timer:',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontWeight': 'bold', 'fontSize': 'large'}
+         ),
+         html.Div(
+             '',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontSize': 'large'},
+             id='buffed_swing_timer'
+         )
+     ]),
+     html.Div([
+         html.Div(
+             'Attack Power:',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontWeight': 'bold', 'fontSize': 'large'}
+         ),
+         html.Div(
+             '',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontSize': 'large'},
+             id='buffed_attack_power'
+         )
+     ]),
+     html.Div([
+         html.Div(
+             'Boss Crit Chance:',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontWeight': 'bold', 'fontSize': 'large'}
+         ),
+         html.Div(
+             '',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontSize': 'large'},
+             id='buffed_crit'
+         )
+     ]),
+     html.Div([
+         html.Div(
+             'Boss Miss Chance:',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontWeight': 'bold', 'fontSize': 'large'}
+         ),
+         html.Div(
+             '',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontSize': 'large'},
+             id='buffed_miss'
+         )
+     ]),
+     html.Div([
+         html.Div(
+             'Mana:',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontWeight': 'bold', 'fontSize': 'large'}
+         ),
+         html.Div(
+             '',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontSize': 'large'},
+             id='buffed_mana'
+         )
+     ]),
+     html.Div([
+         html.Div(
+             'Intellect:',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontWeight': 'bold', 'fontSize': 'large'}
+         ),
+         html.Div(
+             '',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontSize': 'large'},
+             id='buffed_int'
+         )
+     ]),
+     html.Div([
+         html.Div(
+             'Spirit:',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontWeight': 'bold', 'fontSize': 'large'}
+         ),
+         html.Div(
+             '',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontSize': 'large'},
+             id='buffed_spirit'
+         )
+     ]),
+     html.Div([
+         html.Div(
+             'MP5:',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontWeight': 'bold', 'fontSize': 'large'}
+         ),
+         html.Div(
+             '',
+             style={'width': '50%', 'display': 'inline-block',
+                    'fontSize': 'large'},
+             id='buffed_mp5'
+         )
+     ])],
+    width=4, xl=3, style={'marginLeft': '2.5%', 'marginBottom': '2.5%'}
+)
+
+## gear output
+
+#gear_output =
+generate_gear_output = html.Div([
+    dbc.Row(
+        [
+            dbc.Col([
+                html.H4('Gear Sets'),
+                dash_table.DataTable(
+                    data=[],  # df.to_dict('results'),
+                    columns=[
+                        # {'name': i, 'id': i, 'deletable': True} for i in sorted(df.columns)
+                    ],
+                    style_data_conditional=[
+                        {
+                            'if': {'row_index': 'odd'},
+                            'backgroundColor': 'rgb(90, 90, 90)',
+                        },
+                    ],
+                    style_header={
+                        'font_size': '20px',
+                        'font_weight': 'bold',
+                        'backgroundColor': 'rgb(13, 163, 57)',
+                        'color': 'white',
+                        'textAlign': 'left',
+                        'minWidth': 50
+                    },
+                    style_data={
+                        ''
+                        'backgroundColor': 'rgb(50, 50, 50)',
+                        'color': 'white',
+                        'textAlign': 'center',
+                        'minWidth': 50
+                    },
+                    # page_action='custom',
+                    sort_action='custom',
+                    sort_by=[],
+                    id="gear_output"
+                )
+            ],
+                style={'marginLeft': '2.5%', 'marginBottom': '2.5%'},
+                width=4, xl=3,
+                align='center',
+            )
+        ],
+        justify='left',
+    )
+])
+
+
